requirements.txt
kafka-python
pandas 
matplotlib
pymongo
faker